refactor(core): extract AI request timeout helper
